♻️ refactor: update AsyncFileBuffer usage

+++ yutto/utils/time.py
@@ -8,6 +8,6 @@ def get_time_str_by_now():
     return get_time_str_by_stamp(time_stamp_now)
 
 
-def get_time_str_by_stamp(stamp: int):
+def get_time_str_by_stamp(stamp: float):
     local_time = time.localtime(stamp)
     return time.strftime(TIME_FMT, local_time)
